integrate satellite ativity with io
