+++ app/src/main/java/com/ds/avare/MainActivity.java
@@ -69,9 +69,8 @@ public class MainActivity extends TabActivity {
     public static final int tabThreeD = 7;
     public static final int tabChecklist = 8;
     public static final int tabWXB = 9;
-    public static final int tabTools = 10;
-    public static final int tabWnb = 11;
-    public static final int tabIo = 12;
+    public static final int tabWnb = 10;
+    public static final int tabIo = 11;
 
     public void setup() {
         mTextView.setVisibility(View.INVISIBLE);
@@ -117,10 +116,6 @@ public class MainActivity extends TabActivity {
 
         if (0 != (tabItems & (1 << tabWXB))) {
             setupTab(new TextView(this), getString(R.string.WXB), new Intent(this, FaaFileActivity.class), getIntent());
-        }
-
-        if (0 != (tabItems & (1 << tabTools))) {
-            setupTab(new TextView(this), getString(R.string.Tools), new Intent(this, SatelliteActivity.class), getIntent());
         }
 
         if (0 != (tabItems & (1 << tabWnb))) {
