+++ app/src/main/res/layout/layout_play.xml
@@ -38,7 +38,7 @@ Redistribution and use in source and binary forms, with or without modification,
         android:layout_below="@+id/main_image_file"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/File"/>        
+        android:text="@string/FileRead"/>
     <Button
         android:id="@+id/main_button_connect_file"
         android:layout_width="wrap_content"
