+++ app/src/main/res/layout/layout_satellite.xml
