+++ app/src/main/res/xml/preferences.xml
@@ -462,11 +462,6 @@ authors: zkhan, jlmcgraw
                 android:title="@string/WXB" />
             <CheckBoxPreference
                 android:defaultValue="true"
-                android:key="@string/prefTabTools"
-                android:summary="@string/GPSSummary"
-                android:title="@string/Tools" />
-            <CheckBoxPreference
-                android:defaultValue="true"
                 android:key="@string/prefTabWnb"
                 android:summary="@string/WnbSummary"
                 android:title="@string/Wnb" />
